% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donut.R
\name{donut_func}
\alias{donut_func}
\title{Drawing donut}
\usage{
donut_func(df, group = group, length = length, loli_ns = loli_ns, k)
}
\arguments{
\item{df}{Input dataframe}

\item{group}{Subgroup column}

\item{length}{Length of each subgroup (portion in the donut plot)}

\item{loli_ns}{Number of loli sticks in each subgroup}

\item{k}{Label text of each group}
}
\value{
Donut from processed data
}
\description{
Drawing donut plot with lolli sticks given processed data
}
