% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donut.R
\name{generate_sequence}
\alias{generate_sequence}
\title{Helper function to generate coordinates of lolli sticks}
\usage{
generate_sequence(a, b, c)
}
\arguments{
\item{a}{coordinate lowerbound}

\item{b}{coordinate upperbound}

\item{c}{number of lollisticks}
}
\value{
Coordinates of lollisticks given lower and upperbound coordinates
}
\description{
generate coordinates of lolli sticks in donut plot
}
