% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{load_sample_pyclone}
\alias{load_sample_pyclone}
\title{Sample pyclone file for visualization}
\usage{
load_sample_pyclone()
}
\value{
A pyclone sample for drawing donut plot
}
\description{
Return a sample pyclone result, with driver genes annotated
}
\examples{
data <- load_sample_pyclone()
print(head(data))
}
