% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_triangle_plot_function.R
\name{make_full_triangle_plot}
\alias{make_full_triangle_plot}
\title{Base full triangle phylogeny plot}
\usage{
make_full_triangle_plot(
  data,
  x = sample_id,
  y_1 = clonal_driver_count.snv,
  y_2 = clonal_count.cna,
  z_1 = subclonal_driver_count.snv,
  z_2 = subclonal_count.cna,
  show_description = T
)
}
\arguments{
\item{data}{Input dataframe}

\item{x}{Group column}

\item{y_1}{Column represents stem height group A}

\item{y_2}{Column represents stem height 1 of group B}

\item{z_1}{Column represents triangle height/width of group A}

\item{z_2}{Column represents triangle height/width  of group B}

\item{show_description}{Showing detailed numbers on the graph as data description}
}
\value{
A phylogeny plot for ggplot2
}
\description{
Make base phylogeny plot as a full triangle plot. Please note that the plot
can only work on certain number range of the dataframe.
}
\examples{
df <- data.frame(x = 1:10,
                 y1 = sample(1:10, 10, replace = TRUE),
                 y2 = sample(1:10, 10, replace = TRUE),
                 z1 = sample(1:10, 10, replace = TRUE),
                 z2 = sample(1:10, 10, replace = TRUE))
make_full_triangle_plot(df, x = as.factor(x),
                        y_1= y1, y_2= y2,
                        z_1 = z1, z_2 = z2)
}
