% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle_plot_function.R
\name{make_half_triangle_plot}
\alias{make_half_triangle_plot}
\title{Base half-triangle phylogeny plot}
\usage{
make_half_triangle_plot(
  data,
  x = sample_id,
  y_1 = clonal_driver_count.snv,
  y_2 = clonal_count.cna,
  z_1 = subclonal_driver_count.snv,
  z_2 = subclonal_count.cna,
  t_1 = 0.1,
  t_2 = 0.1
)
}
\arguments{
\item{data}{Input dataframe}

\item{x}{Group column}

\item{y_1}{Column represents stem height group A}

\item{y_2}{Column represents stem height 1 of group B}

\item{z_1}{Column represents triangle height of group A}

\item{z_2}{Column represents triangle height  of group B}

\item{t_1}{Column represents triangle width of group A}

\item{t_2}{Column represents triangle width of group B}
}
\value{
A phylogeny plot for ggplot2
}
\description{
Make base phylogeny plot as a half-triangle plot. Please note that the plot
can only work on certain number range of the dataframe.
(trials are encouraged)
}
\examples{
df <- data.frame(x = 1:5,
                 y1 = 1:5,
                 y2 = 1:5,
                 z1 = seq(10, 50, length.out=5),
                 z2 = seq(50, 10, length.out=5))
make_half_triangle_plot(df, x = as.factor(x),
                        y_1= y1, y_2= y2,
                        z_1 = z1, z_2 = z2)
}
