% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donut.R
\name{process_func}
\alias{process_func}
\title{Processing the dataframe}
\usage{
process_func(data, group = group, length = length, loli_ns = loli_ns)
}
\arguments{
\item{data}{Input dataframe}

\item{group}{Subgroup column}

\item{length}{Length of each subgroup (portion in the donut plot)}

\item{loli_ns}{Number of loli sticks in each subgroup}
}
\value{
A data has been processed and summarised
}
\description{
Process the dataframe
}
