% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyclone_plot_function.R
\name{summarise_pyclone_data}
\alias{summarise_pyclone_data}
\title{Summarise pyclone result file}
\usage{
summarise_pyclone_data(data)
}
\arguments{
\item{data}{Pyclone-vi result + isDriver denoting if the mutation is driver.}
}
\value{
Preprocessed pyclone data
}
\description{
Summarise from pyclone data
}
